feat: integration balance
