+++ helper/nominalFormat.js
@@ -1,0 +1,5 @@
+const nominalFormat = (num) => {
+  return Number(num).toLocaleString('id-ID')
+}
+
+export default nominalFormat;
