+++ redux/actions/balance.js
@@ -1,0 +1,5 @@
+import http from '../../helper/http';
+
+const getBalance = (token) => {
+  
+}
