+++ redux/reducers/index.js
@@ -2,14 +2,16 @@ import { combineReducers } from "redux";
 import histories from "./histories";
 import inputTransfer from "./transfer";
 // import registerUser from "./auth";
-import { registerUser, login, phoneList } from "./auth";
+import { registerUser, login } from "./auth";
+import { phoneList, balance } from './profile';
 
 const rootReducers = combineReducers({
   histories,
   inputTransfer,
   registerUser,
   login,
-  phoneList
+  phoneList,
+  balance
 })
 
 export default rootReducers;
